move SyntaxHighlighter component
