+++ apps/explorer/src/components/SyntaxHighlighter/index.tsx
@@ -11,7 +11,7 @@ interface Props {
     language: Language;
 }
 
-export function SyntaxHighlightedCode({ code, language }: Props) {
+export function SyntaxHighlighter({ code, language }: Props) {
     return (
         <section className="px-0 ">
             <div className="overflow-auto whitespace-pre p-2 pl-0 pr-0 font-mono text-sm">
